+++ outwatch/src/main/scala/outwatch/dom/Compat.scala
@@ -1,12 +1,12 @@
 package outwatch
 package dom
 
-import cats.effect.IO
+import cats.effect.SyncIO
 import monix.execution.Scheduler
 import org.scalajs.dom.{ClipboardEvent, DragEvent, KeyboardEvent, MouseEvent}
 import outwatch.dom.helpers.{ChildStreamReceiverBuilder, ChildrenStreamReceiverBuilder}
 
-trait Handlers extends HandlerOps[IO]{
+trait Handlers extends HandlerOps[SyncIO]{
   @deprecated("Use Handler.create[F, MouseEvent] instead", "0.11.0")
   def createMouseHandler()(implicit s: Scheduler) = Handler.create[MouseEvent]
   @deprecated("Use Handler.create[F, KeyboardEvent] instead", "0.11.0")