+++ outwatch/src/main/scala/outwatch/dom/VDomModifier.scala
@@ -1,7 +1,7 @@
 package outwatch.dom
 
 import cats.Monoid
-import cats.effect.IO
+import cats.effect.{IO, SyncIO}
 import monix.execution.Scheduler
 import org.scalajs.dom._
 import outwatch.AsVDomModifier
@@ -76,6 +76,7 @@ final case class NextVDomModifier(modifier: StaticVDomModifier) extends StaticVD
 case object EmptyModifier extends VDomModifier
 final case class CompositeModifier(modifiers: js.Array[_ <: VDomModifier]) extends VDomModifier
 final case class ModifierStreamReceiver(stream: ValueObservable[VDomModifier]) extends VDomModifier
+final case class EffectModifierIO(effect: SyncIO[VDomModifier]) extends VDomModifier
 final case class EffectModifier(effect: IO[VDomModifier]) extends VDomModifier
 final case class SchedulerAction(action: Scheduler => VDomModifier) extends VDomModifier
 final case class StringVNode(text: String) extends VDomModifier